+++ MetaWear/Core/DeviceInformation.swift
@@ -42,12 +42,24 @@ public struct DeviceInformation {
     public let serialNumber: String
     public let firmwareRevision: String
     public let hardwareRevision: String
+    
+    public init(manufacturer: String,
+                modelNumber: String,
+                serialNumber: String,
+                firmwareRevision: String,
+                hardwareRevision: String) {
+        self.manufacturer = manufacturer
+        self.modelNumber = modelNumber
+        self.serialNumber = serialNumber
+        self.firmwareRevision = firmwareRevision
+        self.hardwareRevision = hardwareRevision
+    }
 }
 
 extension MblMwDeviceInformation {
     /// Used to bridge between MetaWearCpp classes and native managed Swift struct
     func convert() -> DeviceInformation {
-        return DeviceInformation(manufacturer:  String(cString: manufacturer),
+        return DeviceInformation(manufacturer: String(cString: manufacturer),
                                  modelNumber: String(cString: model_number),
                                  serialNumber: String(cString: serial_number),
                                  firmwareRevision: String(cString: firmware_revision),