+++ MetaWear/Mocks/MockModule.swift
@@ -43,7 +43,7 @@ struct MessagePayload {
 
 extension Data {
     var message: MessagePayload {
-        return MessagePayload(modId: self[0], regId: self[1], data: suffix(from: 2))
+        return MessagePayload(modId: self[0], regId: self[1], data: Data(suffix(from: 2)))
     }
 }
 
@@ -55,17 +55,17 @@ class MockModule {
     let modId: UInt8
     var handlers: [UInt8: MessageHandler] = [:]
     var entries: [UInt8] = Array(0...0xFF)
-
+    
     init(peripheral: MockPeripheral, modId: UInt8, modImpl: UInt8, modRev: UInt8, extra: Data? = nil) {
         self.peripheral = peripheral
         self.modId = modId
         
-        var header = Data(bytes: [modImpl, modRev])
+        var data = Data(bytes: [modImpl, modRev])
         if let extra = extra {
-            header.append(extra)
+            data.append(extra)
         }
         handleRead(regId: 0) {
-            peripheral.messageSend(modId: $0.regId, regId: $0.regId, notifyEn: true, data: header)
+            peripheral.messageSend(modId: $0.modId, regId: $0.regId, notifyEn: true, data: data)
         }
     }
     
@@ -89,19 +89,137 @@ class MockModule {
         }
     }
     
-    static func logging(peripheral: MockPeripheral) -> MockModule {
-        let extra = Data(bytes: [0x8, 0x0, 0x0, 0x27, 0x10])
-        let modId: UInt8 = 0xB
-        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: 0, modRev: 2, extra: extra)
+    static func mechanicalSwitch(peripheral: MockPeripheral) -> MockModule {
+        let modId: UInt8 = 0x1
+        return MockModule(peripheral: peripheral, modId: modId, modImpl: 0, modRev: 0)
+    }
+    
+    static func led(peripheral: MockPeripheral) -> MockModule {
+        let extra = Data(bytes: [0x3, 0x0])
+        let modId: UInt8 = 0x2
+        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: 0, modRev: 1, extra: extra)
+        return module
+    }
+    
+    static func accelBMI160(peripheral: MockPeripheral) -> MockModule {
+        let extra = Data(bytes: [0x3, 0x0])
+        let modId: UInt8 = 0x3
+        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: 1, modRev: 2, extra: extra)
+        return module
+    }
+    
+    static func iBeacon(peripheral: MockPeripheral) -> MockModule {
+        let modId: UInt8 = 0x7
+        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: 0, modRev: 0)
+        return module
+    }
+    
+    static func genericEntriesModule(peripheral: MockPeripheral,
+                                     modId: UInt8,
+                                     modImpl: UInt8,
+                                     modRev: UInt8,
+                                     extra: Data?,
+                                     createReg: UInt8,
+                                     removeReg: UInt8,
+                                     removeAllReg: UInt8) -> MockModule {
+        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: modImpl, modRev: modRev, extra: extra)
         
-        module.handleWrite(regId: 2) {
-            let lodId = module.entries.remove(at: 0)
-            module.peripheral.messageSend(modId: $0.modId, regId: $0.regId, notifyEn: true, data: Data(bytes: [lodId]))
+        module.handleWrite(regId: createReg) {
+            let eventId = module.entries.remove(at: 0)
+            module.peripheral.messageSend(modId: $0.modId, regId: $0.regId, notifyEn: true, data: Data(bytes: [eventId]))
         }
+        module.handleWrite(regId: removeReg) {
+            let eventId = $0.data[0]
+            let index = module.entries.insertionIndexOf(elem: eventId) { $0 < $1 }
+            module.entries.insert(eventId, at: index)
+        }
+        module.handleWrite(regId: removeAllReg) { _ in
+            module.entries = Array(0...0xFF)
+        }
+        return module
+    }
+    
+    static func dataProcessor(peripheral: MockPeripheral) -> MockModule {
+        return genericEntriesModule(peripheral: peripheral,
+                                    modId: 0x9,
+                                    modImpl: 0,
+                                    modRev: 2,
+                                    extra: Data(bytes: [0x1c]),
+                                    createReg: 2,
+                                    removeReg: 6,
+                                    removeAllReg: 8)
+    }
+    
+    static func event(peripheral: MockPeripheral) -> MockModule {
+        return genericEntriesModule(peripheral: peripheral,
+                                    modId: 0xA,
+                                    modImpl: 0,
+                                    modRev: 0,
+                                    extra: Data(bytes: [0x1c]),
+                                    createReg: 2,
+                                    removeReg: 4,
+                                    removeAllReg: 5)
+    }
+    
+    static func logging(peripheral: MockPeripheral) -> MockModule {
+        let module = genericEntriesModule(peripheral: peripheral,
+                                          modId: 0xB,
+                                          modImpl: 0,
+                                          modRev: 2,
+                                          extra: Data(bytes: [0x8, 0x0, 0x0, 0x27, 0x10]),
+                                          createReg: 0x2,
+                                          removeReg: 0x3,
+                                          removeAllReg: 0xA)
         module.handleRead(regId: 4) {
             module.peripheral.messageSend(modId: $0.modId, regId: $0.regId, notifyEn: true, data: Data(bytes: [0x0, 0x0, 0x0, 0x0, 0x0]))
         }
         return module
     }
+    
+    static func timer(peripheral: MockPeripheral) -> MockModule {
+        return genericEntriesModule(peripheral: peripheral,
+                                    modId: 0xC,
+                                    modImpl: 0,
+                                    modRev: 0,
+                                    extra: Data(bytes: [0x8]),
+                                    createReg: 2,
+                                    removeReg: 5,
+                                    removeAllReg: 8)
+    }
+    
+    static func macro(peripheral: MockPeripheral) -> MockModule {
+        let modId: UInt8 = 0xF
+        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: 0, modRev: 2, extra: Data(bytes: [0x8, 0x7]))
+        return module
+    }
+    
+    static func settings(peripheral: MockPeripheral) -> MockModule {
+        let modId: UInt8 = 0x11
+        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: 0, modRev: 7, extra: Data(bytes: [0x3]))
+        return module
+    }
+    
+    static func testDebug(peripheral: MockPeripheral) -> MockModule {
+        let modId: UInt8 = 0xFE
+        let module = MockModule(peripheral: peripheral, modId: modId, modImpl: 0, modRev: 3)
+        return module
+    }
 }
 
+extension Array {
+    func insertionIndexOf(elem: Element, isOrderedBefore: (Element, Element) -> Bool) -> Int {
+        var lo = 0
+        var hi = self.count - 1
+        while lo <= hi {
+            let mid = (lo + hi)/2
+            if isOrderedBefore(self[mid], elem) {
+                lo = mid + 1
+            } else if isOrderedBefore(elem, self[mid]) {
+                hi = mid - 1
+            } else {
+                return mid // found at position mid
+            }
+        }
+        return lo // not found, would be inserted at position lo
+    }
+}